refactor more utils to common
