+++ src/common/k8s-api/endpoints/cron-job.api.ts
@@ -23,7 +23,7 @@ import moment from "moment";
 import { KubeObject } from "../kube-object";
 import type { IPodContainer } from "./pods.api";
 import { formatDuration } from "../../utils/formatDuration";
-import { autoBind } from "../../../renderer/utils";
+import { autoBind } from "../../utils";
 import { KubeApi } from "../kube-api";
 import type { KubeJsonApiData } from "../kube-json-api";
