+++ src/common/utils/index.ts
@@ -55,6 +55,8 @@ export * from "./toggle-set";
 export * from "./toJS";
 export * from "./type-narrowing";
 export * from "./types";
+export * from "./convertMemory";
+export * from "./convertCpu";
 
 import * as iter from "./iter";
