+++ src/renderer/utils/index.ts
@@ -33,8 +33,8 @@ export * from "./createStorage";
 export * from "./interval";
 export * from "./copyToClipboard";
 export * from "./isReactNode";
-export * from "./convertMemory";
-export * from "./convertCpu";
+export * from "../../common/utils/convertMemory";
+export * from "../../common/utils/convertCpu";
 export * from "./metricUnitsToNumber";
 export * from "./display-booleans";
 export * from "./isMiddleClick";
